requirements.txt
langchain
langchain-core
langchain-groq
langchain-community
pypdf
faiss-cpu
streamlit
sentence-transformers
langgraph
python-dotenv
